refactor(helpers): extract domain helpers and path safety utilities into helpers.js
Moves 11 functions from index.js into helpers.js:

Stateless utilities (exported directly):
- deriveLoglineFromProse, inferCharacterIdsFromProse
- readSupportingNotesForEntity, readEntityMetadata
- resolveBatchTargetScenes

Sync-dir-dependent helpers (createHelpers factory):
- isPathInsideSyncDir, isPathCandidateInsideSyncDir
- resolveOutputDirWithinSync, resolveProjectRoot
- resolveWorldEntityDir, createCanonicalWorldEntity

index.js removes matter, yaml, ReviewBundlePlanError, and
world-entity-templates imports; wires the factory after SYNC_DIR_WRITABLE
is set. toolContext is unchanged — tools receive the same function names.
No behavioral changes.
